requirements.txt
flask
flask_wtf
sqlalchemy
wtforms
flask-login
werkzeug
openpyxl
opencv-python
pyzbar
python-barcode
pytils